прошлась по всем fragment
