+++ app/src/main/java/ru/practicum/android/diploma/presentation/filter/SettingsFilterFragment.kt
@@ -70,4 +70,8 @@ class SettingsFilterFragment: Fragment() {
     }
 
 
+    override fun onDestroy() {
+        super.onDestroy()
+        _binding = null
+    }
 }
