+++ app/src/main/java/ru/practicum/android/diploma/presentation/team/TeamFragment.kt
@@ -20,4 +20,9 @@ class TeamFragment : Fragment() {
         _binding = FragmentTeamBinding.inflate(inflater, container, false)
         return binding.root
     }
+
+    override fun onDestroy() {
+        super.onDestroy()
+        _binding = null
+    }
 }
