+++ app/src/main/res/layout/fragment_choose_area.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".presentation.filter.chooseArea.ChooseAreaFragment">
+    tools:context=".presentation.filter.chooseArea.SelectAreaFragment">
 
     <ImageView
         android:id="@+id/choose_area_back_arrow_imageview"
