+++ app/src/main/res/layout/fragment_choose_workplace.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".presentation.filter.ChooseWorkplaceFragment">
+    tools:context=".presentation.filter.SelectWorkplaceFragment">
 
     <ImageView
         android:id="@+id/settings_back_arrow_imageview"
